Rename parsers/default.js to parsers/index.js.
