+++ lib/options.js
@@ -15,7 +15,7 @@ const defaultOptions = {
   repl: false,
   sourceType: "unambiguous",
   get parse () {
-    return require("./parsers/default.js").parse;
+    return require("./parsers").parse;
   }
 };
